Changes to the TEI Translator
The changes refer to the export of the Zotero item URI, the export
of the short title and the tei-conform export of formatting in titles.

Here is a more detailed description of the changes:

1. the export of the Zotero key
If you enable "Generate XML IDs", an XML ID is generated for
each entry. The current translator uses the author's last name
and the date of publication to generate the XML ID (e.g.
xml:id="McDermott1934"). These values are not stable, they can
change for some reason and a different XML ID can be generated
with the next export (e.g. xml:id="McDermott1934a").

In some cases, a unique and stable value for an item is needed.
We propose to export the item URI for each entry.

The item URI does not replace the XML ID. Zotero item keys can
start with a number but valid values for the attribute @xml:id
have to begin with a letter. Therefore we propose to export the
item URI in addition to the XML ID.

2. export of short title
Content of the field "Short Title" is now exported as the following
<title type="short">Short Title</title>.

3. tei-conform export of formatting in titles
Zotero's tags for formatting (e.g. <i></i> and <sup></sup>) are
replaced by the TEI tags (e.g. <hi rend="italics"></hi> and
<hi rend="sup"></hi>). For example, the text <i>italic text</i>
is exported as follows: &lt;hi rend="italics&gt;italic text&lt;/hi&gt;

Unfortunately, the less-than sign and the greater-than sign are shown
as entities similar to the export result of the current TEI translator
(&lt;i&gt;italic text&lt;/i&gt;).