Enhanced support for new release of Artstor and Shared Shelf Commons (#899)
Provided updated translator for new release of Artstor.

This is refactored code. But the test code does not run in Zotero frame work due to AJAX nature of application. May not work on Chrome and Safari due to DOM detection error in Zotero mainframe.
Added support for external viewer capture